Improve error handling
